refactored fucntion locations
